Improve type-cast in PHYSFS_utf8FromUcs2 Mercury233 proposed changing that in patch #331 - 12 Just looks a bit more "correct" thought it shouldn't have any effect in this case Original cast: u16 to u64 to u32 New cast: u16 to u16 to u32 It shouldn't matter as that cast was likely for another case which we didn't copy from original physfs code. Made some test converting around a complicated utf8 text (https://www.w3.org/2001/06/utf-8-test/UTF-8-demo.html) and as expected no difference.
git-svn-id: https://svn.code.sf.net/p/irrlicht/code/trunk@6787 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/utf8.cpp
@@ -356,7 +356,7 @@ static void PHYSFS_utf8FromUcs4(const u32 *src, char *dst, u64 len)
 
 static void PHYSFS_utf8FromUcs2(const u16 *src, char *dst, u64 len)
 {
-	UTF8FROMTYPE(u64, src, dst, len);
+	UTF8FROMTYPE(u16, src, dst, len);
 } /* PHYSFS_utf8FromUcs2 */
 
 #undef UTF8FROMTYPE
